moving "new" method to default and adding deprecation warnings
