+++ aviary/interface/methods_for_level2.py
@@ -56,7 +56,7 @@ from aviary.subsystems.aerodynamics.aerodynamics_builder import CoreAerodynamics
 from aviary.utils.preprocessors import preprocess_propulsion
 from aviary.utils.merge_variable_metadata import merge_meta_data
 
-from aviary.interface.default_phase_info.two_dof_fiti_copy import add_default_sgm_args
+from aviary.interface.default_phase_info.two_dof_fiti import add_default_sgm_args
 from aviary.mission.gasp_based.idle_descent_estimation import add_descent_estimation_as_submodel
 from aviary.mission.phase_builder_base import PhaseBuilderBase
 
@@ -301,7 +301,7 @@ class AviaryProblem(om.Problem):
                     if self.analysis_scheme is AnalysisScheme.COLLOCATION:
                         from aviary.interface.default_phase_info.two_dof import phase_info
                     elif self.analysis_scheme is AnalysisScheme.SHOOTING:
-                        from aviary.interface.default_phase_info.two_dof_fiti_copy import phase_info, \
+                        from aviary.interface.default_phase_info.two_dof_fiti import phase_info, \
                             phase_info_parameterization
                     phase_info, _ = phase_info_parameterization(
                         phase_info, None, self.aviary_inputs)
