+++ aviary/mission/gasp_based/test/test_idle_descent_estimation.py
@@ -3,7 +3,7 @@ import warnings
 
 import openmdao.api as om
 from aviary.interface.default_phase_info.two_dof_fiti_deprecated import create_2dof_based_descent_phases
-from aviary.interface.default_phase_info.two_dof_fiti_copy import descent_phases, add_default_sgm_args
+from aviary.interface.default_phase_info.two_dof_fiti import descent_phases, add_default_sgm_args
 
 from openmdao.utils.assert_utils import assert_near_equal
